\name{cwhstring-package}
\alias{cwhstring-package}
\docType{package}
\title{
cwh's string manipulation functions
}
\description{
Miscellaneous functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhstring\cr
Version: \tab 4.2.0\cr
Date: \tab 2006-01-20\cr
Depends: \tab R (>= 0.99)\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann/hoffmann.htm\cr
}

Index:
\preformatted{
cap                     Change case of strings
cpos                    Find the position of a substring
datetime                Show date and time in ISO format
delstr                  Delete a substring from a string
dt2str                  Convert time difference to string.
formatFix               Format to a fixed format representation
formula2string          Extract sides of a formula to strings
n22dig                  Show vector or matrix (of 0 <= x <=1) in a
                        compact way:
n2c                     Show absolute values as characters, prepare for
                        plotting
numeric.string          Test string for being a number or made up of
                        digits only.
padding                 Padding a string with justification
pasteInfix              Paste(infix)
pasteRound              Paste rounded values
replacechar             Replace a character in a string by another
splitString             Split a string at the location(s) of a
                        delimiting string, at the single characters of
                        a delimiter
str2formula             Convert back to a formula
strip.whites            Strip white spaces
strmatch                A "shortest unique identifier" match.
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
