\name{cwhstring-internal}
\alias{cwhstring-internal}
\alias{spacC}
\title{The String Package of Christian W. Hoffmann}
\description{Internal String functions and constants}
\author{Christian W. Hoffmann, \email{christian.hoffmann@wsl.ch}}
\details{
  These are not to be called by the user.
}
\keyword{ internal } 